test(executor): split context compaction tests into history and in-flight files
