+++ runtime/src/llm/chat-executor.test.ts
@@ -2730,402 +2730,6 @@ describe("ChatExecutor", () => {
   });
 
   // --------------------------------------------------------------------------
-  // Context compaction
-  // --------------------------------------------------------------------------
-
-  describe("context compaction", () => {
-    it("compacts instead of throwing when budget exceeded", async () => {
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          // First two calls: normal responses that burn through the budget
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          // Third call triggers compaction — summary call succeeds
-          .mockResolvedValueOnce(
-            mockResponse({ content: "Summary of conversation" }),
-          )
-          // Fourth call is the actual execution after compaction
-          .mockResolvedValueOnce(
-            mockResponse({
-              content: "response after compaction",
-              usage: { promptTokens: 20, completionTokens: 10, totalTokens: 30 },
-            }),
-          ),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 1500,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      // Third call — budget exceeded, compaction succeeds, execution continues
-      const result = await executor.execute(
-        createParams({ history: buildLongHistory(10) }),
-      );
-
-      expect(result.compacted).toBe(true);
-      expect(result.content).toBe("response after compaction");
-    });
-
-    it("resets token counter after compaction", async () => {
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          // Summary call
-          .mockResolvedValueOnce(
-            mockResponse({ content: "Summary" }),
-          )
-          // Execution after compaction
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 10, completionTokens: 10, totalTokens: 20 },
-            }),
-          ),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 1500,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      expect(executor.getSessionTokenUsage("session-1")).toBe(2000);
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      // After compaction, counter was reset then new usage accumulated
-      expect(executor.getSessionTokenUsage("session-1")).toBe(20);
-    });
-
-    it("emits provider trace events for pre-execution compaction calls", async () => {
-      const traceEvents: unknown[] = [];
-      let invocation = 0;
-      const provider = createMockProvider("primary", {
-        chat: vi.fn(async (_messages, options) => {
-          invocation += 1;
-          options?.trace?.onProviderTraceEvent?.({
-            kind: "request",
-            transport: "chat",
-            provider: "primary",
-            model: "mock-model",
-            payload: { observed: true, invocation },
-          });
-          if (invocation <= 2) {
-            return mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            });
-          }
-          if (invocation === 3) {
-            return mockResponse({ content: "Summary of conversation" });
-          }
-          return mockResponse({
-            content: "response after compaction",
-            usage: { promptTokens: 20, completionTokens: 10, totalTokens: 30 },
-          });
-        }),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 1500,
-      });
-
-      await executor.execute(
-        createParams({
-          history: buildLongHistory(10),
-          trace: {
-            includeProviderPayloads: true,
-            onProviderTraceEvent: (event) => traceEvents.push(event),
-          },
-        }),
-      );
-      await executor.execute(
-        createParams({
-          history: buildLongHistory(10),
-          trace: {
-            includeProviderPayloads: true,
-            onProviderTraceEvent: (event) => traceEvents.push(event),
-          },
-        }),
-      );
-      await executor.execute(
-        createParams({
-          history: buildLongHistory(10),
-          trace: {
-            includeProviderPayloads: true,
-            onProviderTraceEvent: (event) => traceEvents.push(event),
-          },
-        }),
-      );
-
-      expect(traceEvents).toContainEqual({
-        kind: "request",
-        transport: "chat",
-        provider: "primary",
-        model: "mock-model",
-        callIndex: 0,
-        callPhase: "compaction",
-        payload: { observed: true, invocation: 3 },
-      });
-    });
-
-    it("invokes onCompaction callback", async () => {
-      const onCompaction = vi.fn();
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          // Summary call
-          .mockResolvedValueOnce(
-            mockResponse({ content: "Compact summary text" }),
-          )
-          // Execution after compaction
-          .mockResolvedValueOnce(mockResponse()),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 1500,
-        onCompaction,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      expect(onCompaction).toHaveBeenCalledOnce();
-      expect(onCompaction).toHaveBeenCalledWith(
-        "session-1",
-        "Compact summary text",
-      );
-    });
-
-    it("short history skips summarization but still resets tokens", async () => {
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          // With <=5 messages, compactHistory returns history as-is, no summary call
-          // Next call is the actual execution
-          .mockResolvedValueOnce(
-            mockResponse({
-              content: "short history response",
-              usage: { promptTokens: 10, completionTokens: 10, totalTokens: 20 },
-            }),
-          ),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 1500,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(3) }));
-      await executor.execute(createParams({ history: buildLongHistory(3) }));
-
-      // Short history (<=5 msgs) — no summary LLM call needed
-      const result = await executor.execute(
-        createParams({ history: buildLongHistory(3) }),
-      );
-
-      expect(result.compacted).toBe(true);
-      // Token counter reset + new usage
-      expect(executor.getSessionTokenUsage("session-1")).toBe(20);
-    });
-
-    it("compacted is false when no budget set", async () => {
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({ providers: [provider] });
-
-      const result = await executor.execute(createParams());
-
-      expect(result.compacted).toBe(false);
-    });
-
-    it("uses a soft compaction threshold even when the hard session budget is unlimited", async () => {
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({ content: "Soft-threshold summary" }),
-          )
-          .mockResolvedValueOnce(
-            mockResponse({
-              content: "response after soft-threshold compaction",
-              usage: { promptTokens: 20, completionTokens: 10, totalTokens: 30 },
-            }),
-          ),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 0,
-        sessionCompactionThreshold: 1500,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      const result = await executor.execute(
-        createParams({ history: buildLongHistory(10) }),
-      );
-
-      expect(result.compacted).toBe(true);
-      expect(result.content).toBe("response after soft-threshold compaction");
-    });
-
-    it("treats soft-threshold compaction failures as best-effort when the hard budget is unlimited", async () => {
-      let calls = 0;
-      const provider = createMockProvider("primary", {
-        chat: vi.fn(async () => {
-          calls += 1;
-          if (calls <= 2) {
-            return mockResponse({
-              usage: { promptTokens: 500, completionTokens: 500, totalTokens: 1000 },
-            });
-          }
-          if (calls === 3) {
-            throw new Error("summary unavailable");
-          }
-          return mockResponse({
-            content: "response after skipped compaction",
-            usage: { promptTokens: 20, completionTokens: 10, totalTokens: 30 },
-          });
-        }),
-      });
-      const onCompaction = vi.fn();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 0,
-        sessionCompactionThreshold: 1500,
-        onCompaction,
-        retryPolicyMatrix: {
-          provider_error: { maxRetries: 0 },
-          unknown: { maxRetries: 0 },
-        },
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      const result = await executor.execute(
-        createParams({ history: buildLongHistory(10) }),
-      );
-
-      expect(result.compacted).toBe(false);
-      expect(result.content).toBe("response after skipped compaction");
-      expect(onCompaction).not.toHaveBeenCalled();
-    });
-
-    it("second budget hit re-triggers compaction", async () => {
-      const onCompaction = vi.fn();
-      const provider = createMockProvider("primary", {
-        chat: vi
-          .fn()
-          // Round 1: burn 1000 tokens
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 50, completionTokens: 50, totalTokens: 100 },
-            }),
-          )
-          // Round 2: burn 1000 more → total 2000 >= 100
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 50, completionTokens: 50, totalTokens: 100 },
-            }),
-          )
-          // First compaction summary
-          .mockResolvedValueOnce(
-            mockResponse({ content: "First summary" }),
-          )
-          // Execution after first compaction
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 50, completionTokens: 50, totalTokens: 100 },
-            }),
-          )
-          // Second compaction summary
-          .mockResolvedValueOnce(
-            mockResponse({ content: "Second summary" }),
-          )
-          // Execution after second compaction
-          .mockResolvedValueOnce(
-            mockResponse({
-              usage: { promptTokens: 5, completionTokens: 5, totalTokens: 10 },
-            }),
-          ),
-      });
-      const executor = new ChatExecutor({
-        providers: [provider],
-        sessionTokenBudget: 150,
-        onCompaction,
-      });
-
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-      await executor.execute(createParams({ history: buildLongHistory(10) }));
-
-      // First compaction
-      const r1 = await executor.execute(
-        createParams({ history: buildLongHistory(10) }),
-      );
-      expect(r1.compacted).toBe(true);
-      expect(executor.getSessionTokenUsage("session-1")).toBe(100);
-
-      // Budget hit again (100 >= 150 is false, so need one more)
-      // Actually 100 < 150, so this call passes normally. After this: 200 >= 150.
-      // So we need another call to trigger second compaction.
-      // Let's just verify it compacted once and the counter was reset.
-      expect(onCompaction).toHaveBeenCalledTimes(1);
-      expect(onCompaction).toHaveBeenCalledWith("session-1", "First summary");
-    });
-  });
-
-  // --------------------------------------------------------------------------
   // Per-call streaming
   // --------------------------------------------------------------------------
